requirements/prod.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy-utils
pydantic
psycopg2-binary
